IOMAD: bulk upload should allow for changes to users the manager can't manage. closes #686

+++ blocks/iomad_company_admin/uploaduser.php
@@ -125,6 +125,7 @@ $stremailduplicate          = get_string('useremailduplicate', 'error');
 
 $strinvalidpasswordpolicy   = get_string('invalidpasswordpolicy', 'error');
 $errorstr                   = get_string('error');
+$strcantmanageuser          = get_string('invaliduser', 'block_iomad_company_admin');
 
 $returnurl = $CFG->wwwroot."/blocks/iomad_company_admin/uploaduser.php";
 $bulknurl  = $CFG->wwwroot.'/'.$CFG->admin.'/user/user_bulk.php';
@@ -411,6 +412,11 @@ if ($mform->is_cancelled()) {
                         $deleteerrors++;
                         continue;
                     }
+                    if (!company::check_can_manage($existinguser->id)) {
+                        $upt->track('status', $strcantmanageuser, 'error');
+                        $deleteerrors++;
+                        continue;
+                    }
                     if (delete_user($existinguser)) {
                         $upt->track('status', $struserdeleted);
                         $deletes++;
@@ -446,6 +452,11 @@ if ($mform->is_cancelled()) {
                     $upt->track('id', $olduser->id, 'normal', false);
                     if (is_siteadmin($olduser->id)) {
                         $upt->track('status', $strusernotrenamedadmin, 'error');
+                        $renameerrors++;
+                        continue;
+                    }
+                    if (!company::check_can_manage($olduser->id)) {
+                        $upt->track('status', $strcantmanageuser, 'error');
                         $renameerrors++;
                         continue;
                     }
@@ -507,7 +518,13 @@ if ($mform->is_cancelled()) {
                     $userserrors++;
                     continue;
                 }
-    
+
+                if (!company::check_can_manage($user->id)) {
+                    $upt->track('status', $strcantmanageuser, 'error');
+                    $userserrors++;
+                    continue;
+                }
+
                 if (!empty($updatetype)) {
                     $existinguser->timemodified = time();
                     if (empty($existinguser->timecreated)) {
